Add server command with mock repository
Introduces a complete runnable service for the player data server with graceful
shutdown support and configurable behavior.

Previously, the HTTP server adapter and domain models existed but there was no
executable entrypoint to actually run the service.

This change adds:

- `cmd/serve/main.go` entrypoint with signal handling for SIGINT and SIGTERM,
	enabling graceful shutdown in production deployments

- `service` package providing a `Service` interface and `NewMocker` factory
	that wires together HTTP server and repository components

- Mock player data repository adapter returning hardcoded test data to enable
	development without database dependencies

- Viper configuration library support for externalized config via
	`/etc/player-data-service` directory (overridable via `-config-dir`)

- Server address fallback to `:8080` when not specified in config

The HTTP adapter now takes a `PlayerDataRepository` via constructor injection,
separating server logic from data access and allowing mock‑based testing or
real implementations. However, because the project’s timeline is overrun, we
won’t add those tests right now.

The Vector2 coordinate type changed from float32 to float64 to match standard
JSON number precision and prevent rounding errors in position calculations.

+++ internal/domain/player/player.go
@@ -13,5 +13,5 @@ type Data struct {
 }
 
 type Vector2 struct {
-	X, Y float32
+	X, Y float64
 }
